+++ CoinUtils/src/CoinFileIO.cpp
@@ -64,7 +64,7 @@ public:
 
   virtual int read (void *buffer, int size)
   {
-    return fread (buffer, 1, size, f_);
+    return (int)fread (buffer, 1, size, f_);
   }
 
   virtual char *gets (char *buffer, int size)
@@ -109,7 +109,7 @@ public:
     // First consume data from buffer if available.
     if (dataStart_ < dataEnd_)
       {
-	int amount = dataEnd_ - dataStart_;
+	int amount = (int)(dataEnd_ - dataStart_);
 	if (amount > size)
 	  amount = size;
 
@@ -147,7 +147,7 @@ public:
 	if (dataStart_ == dataEnd_)
 	  {
 	    dataStart_ = dataEnd_ = &dataBuffer_[0];
-	    int count = readRaw (dataStart_, dataBuffer_.size ());
+	    size_t count = readRaw (dataStart_, (int)dataBuffer_.size ());
 
 	    // at EOF?
 	    if (count <= 0) 
@@ -303,7 +303,7 @@ CoinFileInput *CoinFileInput::create (const std::string &fileName)
 {
   // first try to open file, and read first bytes 
   unsigned char header[4];
-  int count ; // So stdin will be plain file
+  size_t count ; // So stdin will be plain file
   if (fileName!="stdin") {
     FILE *f = fopen (fileName.c_str (), "r");
 
@@ -389,7 +389,7 @@ public:
 
   virtual int write (const void *buffer, int size)
   {
-    return fwrite (buffer, 1, size, f_);
+    return (int)fwrite (buffer, 1, size, f_);
   }
 
   // we have something better than the default implementation
@@ -567,7 +567,7 @@ CoinFileOutput::~CoinFileOutput ()
 
 bool CoinFileOutput::puts (const char *s)
 {
-  int len = strlen (s);
+  int len = (int)strlen (s);
   if (len == 0)
     return true;
 
@@ -587,7 +587,7 @@ bool fileAbsPath (const std::string &path)
   // If the first two chars are drive designators then treat it as absolute
   // path (noone in their right mind would create a file named 'Z:' on unix,
   // right?...)
-  const int len = path.length();
+  const size_t len = path.length();
   if (len >= 2 && path[1] == ':') {
     const char ch = path[0];
     if (('a' <= ch && ch <= 'z') || ('A' <= ch && ch <= 'Z')) {
